feat(user): add UserService for user management and refactor today command
Extract user creation/update logic into UserService for better reusability. Move today command implementation to a separate module for better code organization. Update message handler to use the new UserService.
