+++ prisma/schema.prisma
@@ -3,7 +3,6 @@
 
 generator client {
   provider = "prisma-client-js"
-  output = "./src/generated"
 }
 
 datasource db {
