+++ src/services/prisma.ts
@@ -1,3 +1,3 @@
-import { PrismaClient } from "../../prisma/src/generated";
+import { PrismaClient } from "@prisma/client";
 
 export const prisma = new PrismaClient();
